fix() : Rename methods to appropriately
